ai/extract_keyframe/requirements.txt
onedrivedownloader
pygments
pillow
requests
tqdm
pandas
opencv-python
open_clip_torch
